requirements.txt
anyio==3.2.1
argcomplete==1.12.3
asgiref==3.3.4
attrs==21.2.0
boto3==1.17.95
botocore==1.20.95
certifi==2021.5.30
cffi==1.14.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
durationpy==0.5
Faker==8.9.1
Flask==2.0.1
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
h11==0.12.0
hjson==3.0.2
httpcore==0.13.6
httpx==0.18.2
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==2.0.1
packaging==20.9
pep517==0.10.0
pip-tools==6.2.0
placebo==0.9.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==3.0.6
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
s3transfer==0.4.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.1
troposphere==2.7.1
urllib3==1.26.5
Werkzeug==2.0.1
wsgi-request-logger==0.4.6
yandexcloud==0.92.0
-e git+https://github.com/e-kor/yappa.git@b5561e541fd3b16987782c87d60bdd9c6dda9fdc#egg=yappa
